+++ Team_1/Team_1/Monster_B.cpp
@@ -62,10 +62,10 @@ void CMonster_B::Render(HDC _hDC)
 	MoveToEx(_hDC, m_tRC.left + (22 * m_iReverse), m_tRC.bottom - 3, nullptr);
 	LineTo(_hDC, m_tRC.right, m_tRC.bottom);
 
-	MoveToEx(_hDC, m_tRC.left + (22 * m_iReverse), m_tInfo.fY, nullptr);
-	LineTo(_hDC, m_tRC.left + (3 * m_iReverse), m_tInfo.fY);
-	MoveToEx(_hDC, m_tRC.left + (22 * m_iReverse), m_tInfo.fY + 5, nullptr);
-	LineTo(_hDC, m_tRC.left + (8 * m_iReverse), m_tInfo.fY);
+	MoveToEx(_hDC, m_tRC.left + (22 * m_iReverse), (int)m_tInfo.fY, nullptr);
+	LineTo(_hDC, m_tRC.left + (3 * m_iReverse), (int)m_tInfo.fY);
+	MoveToEx(_hDC, m_tRC.left + (22 * m_iReverse), (int)m_tInfo.fY + 5, nullptr);
+	LineTo(_hDC, m_tRC.left + (8 * m_iReverse), (int)m_tInfo.fY);
 }
 
 void CMonster_B::Release(void)
